+++ src/index.ts
@@ -14,8 +14,7 @@ export * from './plugins';
 //TODO: use new syntax
 //TODO: new matrix grid
 //TODO: directive for attaching external plugins, new type that allows obtaining or using export as ref
-//TODO: rework all pagings, styles and use it as component itself
+//TODO: rework all pagings, styles and use it as component itself, paging initialization rework
 //TODO: styles in datetime style
-//TODO: initialize boolean force
 //TODO: items per page overwrite, not merge
 //TODO: remove $any when possible