+++ src/pipes/index.ts
@@ -1,1 +1,2 @@
+export * from './readValue/readValue.pipe';
 export * from './resolveForwardRef/resolveForwardRef.pipe';