+++ src/plugins/contentRenderer/bodyHeader/cssDivs/body/cssDivsBodyContentRenderer.component.ts
@@ -7,6 +7,7 @@ import {TableGridMetadata} from '../../../../../components/tableGridMetadataGath
 import {TableGridColumn} from '../../../../../interfaces';
 import {BODY_CONTENT_RENDERER_OPTIONS, GRID_PLUGIN_INSTANCES} from '../../../../../misc/tokens';
 import {GridPluginInstances} from '../../../../../misc/types';
+import {ReadValueSAPipe} from '../../../../../pipes';
 
 /**
  * Default options for 'CssDivsBodyContentRendererComponent'
@@ -33,6 +34,7 @@ const defaultOptions: CssDivsBodyContentRendererOptions =
     imports:
     [
         CommonModule,
+        ReadValueSAPipe,
     ],
     changeDetection: ChangeDetectionStrategy.OnPush,
 })